Remove KailleraGame interface
