+++ emulinker/src/main/java/org/emulinker/kaillera/model/KailleraServer.kt
@@ -34,7 +34,6 @@ import org.emulinker.kaillera.model.event.UserQuitEvent
 import org.emulinker.kaillera.model.exception.*
 import org.emulinker.kaillera.model.impl.AutoFireDetector
 import org.emulinker.kaillera.model.impl.AutoFireDetectorFactory
-import org.emulinker.kaillera.model.impl.KailleraGameImpl
 import org.emulinker.kaillera.model.impl.Trivia
 import org.emulinker.kaillera.pico.AppModule
 import org.emulinker.kaillera.pico.CompiledFlags
@@ -73,7 +72,7 @@ class KailleraServer(
 
   val usersMap: MutableMap<Int, KailleraUser> = ConcurrentHashMap(flags.maxUsers)
 
-  var gamesMap: MutableMap<Int, KailleraGameImpl> = ConcurrentHashMap(flags.maxGames)
+  var gamesMap: MutableMap<Int, KailleraGame> = ConcurrentHashMap(flags.maxGames)
 
   var trivia: Trivia? = null
 
@@ -651,9 +650,9 @@ class KailleraServer(
         )
       }
     }
-    val game: KailleraGameImpl?
+    val game: KailleraGame?
     val gameID = getNextGameID()
-    game = KailleraGameImpl(gameID, romName, owner = user, this, flags.gameBufferSize, flags, clock)
+    game = KailleraGame(gameID, romName, owner = user, this, flags.gameBufferSize, flags, clock)
     gamesMap[gameID] = game
     addEvent(GameCreatedEvent(this, game))
     logger.atInfo().log("%s created: %s: %s", user, game, game.romName)
@@ -700,7 +699,7 @@ class KailleraServer(
       logger.atSevere().log("%s close %s failed: not in list: %s", user, game, game)
       return
     }
-    (game as KailleraGameImpl).close(user)
+    game.close(user)
     gamesMap.remove(game.id)
     logger.atInfo().log("%s closed: %s", user, game)
     addEvent(GameClosedEvent(this, game))
