+++ emulinker/src/main/java/org/emulinker/kaillera/model/KailleraUser.kt
@@ -22,7 +22,6 @@ import org.emulinker.kaillera.model.event.StopFlagEvent
 import org.emulinker.kaillera.model.event.UserQuitEvent
 import org.emulinker.kaillera.model.event.UserQuitGameEvent
 import org.emulinker.kaillera.model.exception.*
-import org.emulinker.kaillera.model.impl.KailleraGameImpl
 import org.emulinker.util.CircularVariableSizeByteArrayBuffer
 import org.emulinker.util.EmuLang
 import org.emulinker.util.TimeOffsetCache
@@ -209,7 +208,7 @@ class KailleraUser(
     lastKeepAlive = clock.now()
   }
 
-  var game: KailleraGameImpl? = null
+  var game: KailleraGame? = null
     set(value) {
       if (value == null) {
         playerNumber = -1
@@ -343,7 +342,7 @@ class KailleraUser(
     //
     // }
     playerNumber = game.join(this)
-    this.game = game as KailleraGameImpl
+    this.game = game
     gameDataErrorTime = -1
     return game
   }
